fix(Markdown::Handler): improve math block handling
The math block handling logic was flawed, especially when
dealing with inline math expressions or math delimiters
appearing on the same line as other text. This commit
addresses these issues by:

- Correctly identifying the start and end of math blocks,
  even when they are not on separate lines.
- Properly handling inline math expressions enclosed in
  `$$` or `\[` and `\]`.
- Ensuring that the text surrounding math delimiters is
  correctly processed and added to the block.

feat(mdwrap): add wrap_sentence option to enable/disable sentence wrapping

This commit introduces a new option, `wrap_sentence`, to
the `mdwrap` script. This option allows users to control
whether or not sentences should be wrapped at the end of
a line.

- Added `wrap_sentence` option to `GetOptions`.
- Set the default value of `wrap_sentence` to 1 (enabled).